created ScormEventTest and refactored camelcase
