+++ tests/EventTest.php
@@ -128,25 +128,6 @@ abstract class EventTest extends PhpUnitTestCase {
         ];
     }
 
-    protected function constructScormTracking() {
-        return [
-            'scorm_score_raw' => 100,
-            'scorm_score_min' => 0,
-            'scorm_score_scaled' => 1,
-            'scorm_score_max' => 100,
-            'scorm_status' => 'completed',
-        ];
-    }
-    protected function constructScormScoes() {
-        return [
-            'scorm_scoes_id' =>  1,
-            'scorm_scoes_url' =>  'http://www.example.com/module_url',
-            'scorm_scoes_type' => static::$xapi_type.'sco',
-            'scorm_scoes_name' => 'Sco name',
-            'scorm_scoes_description' => 'Sco Description',
-        ];
-    }
-
     protected function assertOutput($input, $output) {
         $this->assertUser($input, $output['actor'], 'user');
         $this->assertObject('app', $input, $output['context']['contextActivities']['grouping'][0]);
